+++ tests/EncDotNet.S100.Pipelines.Tests/DatasetPipelineFactoryRegistryDetectionTests.cs
@@ -1,5 +1,6 @@
 using EncDotNet.S100.Datasets.Pipelines;
 using EncDotNet.S100.Datasets.S57;
+using EncDotNet.S100.TestSupport;
 
 namespace EncDotNet.S100.Pipelines.Tests;